Refactoring: move duplicated code to a base class and common utils.
